Fixed bugs, now shows list, and correctly downloads and inserts RSS
ListFragment was being shown with width of 0dp, now is same as parent.

RSS downloads and parses content properly.  Uses characters(ch, start,
length) start and length params to append(ch, start, length).  was using
content.append(ch) which was putting characters from previous elements
into the current elements content.  Was causing description from
previous elements to show up in the link data.

Now before any new inserts, deletes all old content from the database. 
This is just a temporary solution.

Created a custom MyCursorAdapter instead of SimpleCursorAdapter.  This
was really an attempt to debug why the list wasn't showing, before it
was found the ListFragment was being displayed with a width of 0dp.  Am
keeping it for the moment.

Added the ability to delete all data from database, so that we can do a
mass delete before downloading and adding new content.

Am slowly working towards a Dependency Injection model, using Factories
to create the components.  Have done this for Items so far.